added functionality for no cache timeout

+++ src/cache.rs
@@ -34,44 +34,47 @@ impl Cache {
     pub async fn new(data_timeout: Duration, callback: Option<Callback>) -> Self {
         let hmap: DB = Arc::new(Mutex::new(HashMap::new()));
 
-        // TODO: Negative Duration should mean no timeout
+        // TODO: zero duration should mean no timeout
 
         // For each new cache, spawn a loop which erases all data when it excedes the deadlines.
-        let h = hmap.clone();
-        tokio::spawn(async move {
-            // Vec for storing the dead (timed out) keys from the hashmap.
-            // This looks really weird since we allocate the vec at new-time
-            // when the hashmap does not have any elements in it...?
-            let mut dead: Vec<String> = Vec::with_capacity(h.lock().unwrap().len());
+        // Only do this when the duration is greater than 0, 0 should mean no timeout.
+        if !data_timeout.is_zero() {
+            let h = hmap.clone();
+            tokio::spawn(async move {
+                // Vec for storing the dead (timed out) keys from the hashmap.
+                // This looks really weird since we allocate the vec at new-time
+                // when the hashmap does not have any elements in it...?
+                let mut dead: Vec<String> = Vec::with_capacity(h.lock().unwrap().len());
 
-            loop {
-                // Let the function loop on a interval.
-                tokio::time::sleep(Duration::new(0, 1)).await;
+                loop {
+                    // Let the function loop on a interval.
+                    tokio::time::sleep(Duration::new(0, 1)).await;
 
-                let mut map = h.lock().unwrap();
+                    let mut map = h.lock().unwrap();
 
-                // Check which pieces of cache data overstayed their welcome (have a expired deadline).
-                // add them to the dead vector for easy removal, dont this way to appease the
-                // borrow checker...
-                for (k, data) in map.iter() {
-                    if data.creation.elapsed() > data_timeout {
-                        dead.push(k.to_string());
-                    }
-                }
-
-                // Remove all the now dead (timed out) pieces of data from the hashmap.
-                for key in &dead {
-                    if let Some(f) = callback {
-                        f(key.clone());
+                    // Check which pieces of cache data overstayed their welcome (have a expired deadline).
+                    // add them to the dead vector for easy removal, dont this way to appease the
+                    // borrow checker...
+                    for (k, data) in map.iter() {
+                        if data.creation.elapsed() > data_timeout {
+                            dead.push(k.to_string());
+                        }
                     }
 
-                    map.remove_entry(key);
-                }
+                    // Remove all the now dead (timed out) pieces of data from the hashmap.
+                    for key in &dead {
+                        if let Some(f) = callback {
+                            f(key.clone());
+                        }
 
-                // Reset the array for storing new dead data.
-                dead.clear();
-            }
-        });
+                        map.remove_entry(key);
+                    }
+
+                    // Reset the array for storing new dead data.
+                    dead.clear();
+                }
+            });
+        }
 
         Cache { hmap }
     }
